Rename support/virtualDom to support/d
