+++ README.md
@@ -418,71 +418,10 @@ callListener(node, 'onClick', {
 });
 ```
 
-### support/loadJsdom
-
-`loadJsdom` is a module which will attempt to load `jsdom` in environments where there appears to be no global `document` object
-(e.g. NodeJS).  If it detects `jsdom` needs to be loaded, it will create a global `document` and `window` as well as provide a
-couple of key shims/polyfills to support certain feature detections needed by Dojo 2.
-
-The module's default export is a reference to `document`, either the created one, or the one that is already there.  It will
-essentially be a "noop" if it is running in a browser environment, so it is safe to load without knowing what sort of environment
-you are running in.
-
-Typical usage would be to load the module before starting any client unit tests that need a browser environment:
-
-```typescript
-import '@dojo/intern-helper/support/loadJsdom';
-import 'testModule';
-```
-
-### support/sendEvent()
-
-Dispatch an event to a specified DOM element.  The first argument is the target, the second argument is the type of the event to
-dispatch to the target.  The third is an optional object of additional options:
-
-|Option|Description|
-|------|-----------|
-|eventClass|A string that matches the class of event to use (e.g. `MouseEvent`).  By default, `CustomEvent` is used.|
-|eventInit|Any properties that should be part of initialising the event.  Note that `bubbles` and `cancelable` are `true` by default, which is different then if you were creating events directly.|
-|selector|A string selector to be applied to the target element.|
-
-An example of clicking on a button:
-
-```typescript
-const button = document.createElement('button');
-document.body.appendChild(button);
-sendEvent(button, 'click');
-```
-
-An example of swiping right on a `div`:
-
-```typescript
-const div = document.createElement('div');
-document.body.appendChild(div);
-
-sendEvent(div, 'touchstart', {
-    eventInit: {
-        changedTouches: [ { screenX: 50 } ]
-    }
-});
-
-sendEvent(div, 'touchmove', {
-    eventInit: {
-        changedTouches: [ { screenX: 150 } ]
-    }
-});
-
-sendEvent(div, 'touchend', {
-    eventInit: {
-        changedTouches: [ { screenX: 150 } ]
-    }
-});
-```
-
-### support/virtualDom
+### support/d
 
 In testing expected virtual DOM, it can be overly verbose to regenerate your virtual DOM every time you have changed the conditions
-that might effect the render.  Therefore the `support/virtualDom` module contains some helper functions which can be used to manipulate
+that might effect the render.  Therefore the `support/d` module contains some helper functions which can be used to manipulate
 virtual DOM once it has been generated by the `v()` and `w()` virtual DOM functions.
 
 #### assignChildProperties()
@@ -614,6 +553,67 @@ const expected = v('div', {
 assignProperties(expected, {
     classes: widget.classes(css.root, css.open),
     onclick: widget.listener
+});
+```
+
+### support/loadJsdom
+
+`loadJsdom` is a module which will attempt to load `jsdom` in environments where there appears to be no global `document` object
+(e.g. NodeJS).  If it detects `jsdom` needs to be loaded, it will create a global `document` and `window` as well as provide a
+couple of key shims/polyfills to support certain feature detections needed by Dojo 2.
+
+The module's default export is a reference to `document`, either the created one, or the one that is already there.  It will
+essentially be a "noop" if it is running in a browser environment, so it is safe to load without knowing what sort of environment
+you are running in.
+
+Typical usage would be to load the module before starting any client unit tests that need a browser environment:
+
+```typescript
+import '@dojo/intern-helper/support/loadJsdom';
+import 'testModule';
+```
+
+### support/sendEvent()
+
+Dispatch an event to a specified DOM element.  The first argument is the target, the second argument is the type of the event to
+dispatch to the target.  The third is an optional object of additional options:
+
+|Option|Description|
+|------|-----------|
+|eventClass|A string that matches the class of event to use (e.g. `MouseEvent`).  By default, `CustomEvent` is used.|
+|eventInit|Any properties that should be part of initialising the event.  Note that `bubbles` and `cancelable` are `true` by default, which is different then if you were creating events directly.|
+|selector|A string selector to be applied to the target element.|
+
+An example of clicking on a button:
+
+```typescript
+const button = document.createElement('button');
+document.body.appendChild(button);
+sendEvent(button, 'click');
+```
+
+An example of swiping right on a `div`:
+
+```typescript
+const div = document.createElement('div');
+document.body.appendChild(div);
+
+sendEvent(div, 'touchstart', {
+    eventInit: {
+        changedTouches: [ { screenX: 50 } ]
+    }
+});
+
+sendEvent(div, 'touchmove', {
+    eventInit: {
+        changedTouches: [ { screenX: 150 } ]
+    }
+});
+
+sendEvent(div, 'touchend', {
+    eventInit: {
+        changedTouches: [ { screenX: 150 } ]
+    }
 });
 ```
