+++ tests/unit/support/all.ts
@@ -2,6 +2,6 @@ import './AssertionError';
 import './assertRender';
 import './callListener';
 import './compare';
+import './d';
 import './loadJsdom';
 import './sendEvent';
-import './virtualDom';
